UI-28:  Made changes to unit test case in order to correct compile erros involving creditor/customer balance logic changes.

+++ src/test/java/org/rmt2/messages/accounting/receipts/CustomerCashPaymentUpdateRequestBuilderTest.java
@@ -94,7 +94,7 @@ public class CustomerCashPaymentUpdateRequestBuilderTest {
                 .withAccountNo("ACCT-NO-8888")
                 .withCreditLimit(1234.55)
                 .withAcctDescription("ACCOUNT DESCRIPTION")
-                .withBalance(50000)
+                .withBalance((double) 50000)
                 .withActive(1).build();
         
         xt.setCustomer(customer);
