+++ src/test/java/org/rmt2/messages/accounting/subsidiary/CustomerQueryRequestBuilderTest.java
@@ -65,7 +65,7 @@ public class CustomerQueryRequestBuilderTest {
                 .withAccountNo("ACCT-NO-8888")
                 .withCreditLimit(1234.55)
                 .withAcctDescription("ACCOUNT DESCRIPTION")
-                .withBalance(50000)
+                .withBalance((double) 50000)
                 .withActive(1).build();
                 
         CustomerCriteriaType criteria = fact.createCustomerCriteriaType();
